refactor: fix symlink module to use io_uring operations with DirectoryFd
- Fix symlink module to use io_uring::opcode::SymlinkAt instead of nix
- Implement SymlinkOp struct with proper OpCode trait implementation
- Use types::Fd wrapper for DirectoryFd in io_uring operations
- Update symlink tests to work with io_uring implementation
- Fix integration tests to use create_symlink_at_dirfd consistently
- Update path assertions in tests to match relative symlink targets
- Remove unused nix::unistd import from symlink module
- All symlink tests now passing with io_uring operations

The symlink module now correctly uses io_uring operations as specified,
with secure *at variants using DirectoryFd for enhanced security.

+++ crates/compio-fs-extended/src/device.rs
@@ -61,7 +61,7 @@ pub async fn create_special_file_at_path(path: &Path, mode: u32, dev: u64) -> Re
             &path,
             stat::SFlag::from_bits_truncate(mode),
             stat::Mode::from_bits_truncate(mode & 0o777),
-            dev as u64,
+            dev,
         )
         .map_err(|e| device_error(&format!("mknod failed: {}", e)))
     })
